Apr 21, 2025, 10:10 PM

+++ interpreter/interpreter.py
@@ -3,8 +3,8 @@
 expression = input("Expression: ").strip()
 
 x, y, z = expression.split(" ")
-x = float(x)
-z = float(z)
+x = int(x)
+z = int(z)
 match y:
     case "+":
         print(round(x + z, 1))
